check if roles existing before inserting roles in seed file

+++ src/db/seed/insert-roles.ts
@@ -2,6 +2,11 @@ import db from "../db";
 import { roles, type InsertRole } from "../index";
 
 async function insertRoles(): Promise<object[]> {
+  const currentRoles = await db.select().from(roles);
+  if (currentRoles.length > 0) {
+    return currentRoles;
+  }
+
   await db.delete(roles);
 
   const newRoles: InsertRole[] = [
